CM 18.2: correccion de querys

+++ app/src/main/java/com/passioagogo/market/domain/repository/PADBRepository.kt
@@ -6,7 +6,7 @@ import com.passioagogo.market.data.local.entity.PAProductEntity
 interface PADBRepository {
     fun getAllProduct() : Flow<List<PAProductEntity>>
     fun searchProductsByName(searchTerm: String): Flow<List<PAProductEntity>>
-    suspend fun insertProduct(product: PAProductEntity): Flow<Int>
+    suspend fun insertProduct(product: PAProductEntity): Flow<Long>
     fun updateProduct(product: PAProductEntity): Flow<Int>
     fun deleteProduct(id: Int): Flow<Int>
